+++ src/main/java/com/finanzas/breadcredit/repository/InstallmentRepository.java
@@ -3,5 +3,5 @@ package com.finanzas.breadcredit.repository;
 import com.finanzas.breadcredit.entity.Installment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstallmentRepository extends JpaRepository<Installment, Integer> {
+public interface InstallmentRepository extends JpaRepository<Installment, Long> {
 }
